# openapi.yml
openapi: 3.0.3
info:
  title: Easy!Appointments API
  version: 1.0.0
  description: These are the OpenAPI specs that describe the REST API of Easy!Appointments.
  termsOfService: https://easyappointments.org/terms-and-conditions
  contact:
    name: Easy!Appointments
    url: https://easyappointments.org
    email: info@easyappointments.org
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: https://demo.easyappointments.org/index.php/api/v1/
  - url: http://localhost/index.php/api/v1/
...
paths:
...
  /appointments:
    get:
      tags:
        - appointments
      summary: Get all appointments
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: length
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: fields
          in: query
          schema:
            type: string
        - name: with
          in: query
          schema:
            type: string
        - name: aggregates
          in: query
          schema:
            type: string
          deprecated: true
        - name: date
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: till
          in: query
          schema:
            type: string
        - name: serviceId
          in: query
          schema:
            type: integer
        - name: providerId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentRecord'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    post:
      tags:
        - appointments
      summary: Create an appointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentPayload'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRecord'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: body
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
  /appointments/{appointmentId}:
    get:
      tags:
        - appointments
      summary: Get an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRecord'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    put:
      tags:
        - appointments
      summary: Update an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentPayload'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRecord'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: body
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    delete:
      tags:
        - appointments
      summary: Delete an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
...
  /feedback:
    get:
      tags:
        - feedback
      summary: Get all feedback
      parameters:
        - name: providerId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: appointmentId
          in: query
          schema:
            type: integer
        - name: approved
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: length
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
        - name: fields
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackRecord'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    post:
      tags:
        - feedback
      summary: Create feedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackPayload'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackRecord'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
...
components:
  schemas:
...
    AppointmentRecord:
      type: object
      properties:
        id:
          type: integer
        book:
          type: string
        start:
          type: string
        end:
          type: string
        hash:
          type: string
        location:
          type: string
        color:
          type: string
        status:
          type: string
        notes:
          type: string
        customerId:
          type: integer
        providerId:
          type: integer
        serviceId:
          type: integer
        googleCalendarId:
          type: string
        caldavCalendarId:
          type: string
        paymentStatus:
          type: string
          description: Payment status (pending, paid)
        paymentMethod:
          type: string
          description: Payment method (online, offline)
        stripeSessionId:
          type: string
          description: Stripe checkout session ID
        stripePaymentIntentId:
          type: string
          description: Stripe payment intent ID
        paymentAmount:
          type: number
          format: float
          description: Payment amount
        paymentCurrency:
          type: string
          description: Payment currency (e.g. AED, USD)
        refundAmount:
          type: number
          format: float
          description: Refund amount (if refunded)
        refundStatus:
          type: string
          description: Refund status (none, pending, completed, failed)
          enum:
            - none
            - pending
            - completed
            - failed
        refundReason:
          type: string
          description: Reason for refund
      example:
        id: 1
        book: '2021-01-01 12:00:00'
        start: '2021-01-01 17:00:00'
        end: '2021-01-01 18:00:00'
        hash: apTWVbSvBJXR
        location: Test Street 1A, 12345 Some State, Some Place
        color: '#123456'
        status: Booked
        notes: This is a test appointment.
        customerId: 5
        providerId: 2
        serviceId: 6
        googleCalendarId: null
        caldavCalendarId: null
        paymentStatus: paid
        paymentMethod: online
        stripeSessionId: cs_test_abc123
        stripePaymentIntentId: pi_abc123
        paymentAmount: 250.00
        paymentCurrency: AED
        refundAmount: null
        refundStatus: none
        refundReason: null
    AppointmentPayload:
      type: object
      properties:
        start:
          type: string
        end:
          type: string
        location:
          type: string
        color:
          type: string
        status:
          type: string
        notes:
          type: string
        customerId:
          type: integer
        providerId:
          type: integer
        serviceId:
          type: integer
        stripeSessionId:
          type: string
          description: Stripe checkout session ID (required for online payment). Either this or paymentMethod=offline must be provided when the service has a price > 0.
        paymentMethod:
          type: string
          description: Payment method. Use "offline" for pay-at-clinic. Required when stripeSessionId is not provided and the service has a price > 0.
          enum:
            - online
            - offline
      example:
        start: '2021-01-01 17:00:00'
        end: '2021-01-01 18:00:00'
        location: Test Street 1A, 12345 Some State, Some Place
        color: '#123456'
        status: Booked
        notes: This is a test appointment.
        customerId: 5
        providerId: 2
        serviceId: 6
        stripeSessionId: cs_test_abc123
...
    FeedbackRecord:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        customerId:
          type: integer
        providerId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedbackText:
          type: string
        feedbackCategory:
          type: string
          enum:
            - service
            - doctor
            - facility
            - overall
        isApproved:
          type: boolean
        submittedDate:
          type: string
          format: date-time
        approvedDate:
          type: string
          format: date-time
      example:
        id: 1
        appointmentId: 10
        customerId: 5
        providerId: 2
        rating: 4
        feedbackText: Great service
        feedbackCategory: overall
        isApproved: false
        submittedDate: '2026-02-17T10:00:00'
        approvedDate: null
    FeedbackPayload:
      type: object
      required:
        - appointmentId
        - customerId
        - providerId
        - rating
      properties:
        appointmentId:
          type: integer
        customerId:
          type: integer
        providerId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedbackText:
          type: string
        feedbackCategory:
          type: string
          enum:
            - service
            - doctor
            - facility
            - overall
          default: overall
      example:
        appointmentId: 10
        customerId: 5
        providerId: 2
        rating: 4
        feedbackText: Great service
        feedbackCategory: overall
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      example:
        success: false
        message: Test error message with debug info.
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: >-
        You can generate the API key from the settings page to authorize with
        Bearer Token.
    BasicAuth:
      type: http
      scheme: basic
      description: >-
        You can use the credentials of an admin user to authorize with Basic
        Auth.